+++ src/test/java/courier/CreateCourierTest.java
@@ -1,4 +1,4 @@
-package couriertest;
+package courier;
 
 
 import client.Courier;
