+++ src/test/java/courier/LoginCourierTest.java
@@ -1,4 +1,4 @@
-package couriertest;
+package courier;
 
 import client.Courier;
 import generator.CreateCourierGenerator;
